+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.5.0-dev
+* Deprecated `AnalysisSession.getUnitElement()`.
+  Use `AnalysisSession.getUnitElement2()` instead.
+
 ## 1.4.0
 * Deprecated `TypeProvider.nonSubtypableClasses`.
   Use `TypeProvider.isNonSubtypableClass` instead.